backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
pydantic
twilio